+++ services/property-svc/internal/property-set/projections/spiceDBProjection/set_spicedb_projection.go
@@ -1,4 +1,4 @@
-package propertySetSpiceDBProjection
+package spiceDBProjection
 
 import (
 	"context"
