+++ services/property-svc/internal/property/projections/spiceDBProjection/property_spicedb_projection.go
@@ -1,4 +1,4 @@
-package propertySpiceDBProjection
+package spiceDBProjection
 
 import (
 	"context"
